Hold a map id in 32 bits everywhere it is stored
The other half of the widening. A vessel's deck is a map minted above a
million, and three places still kept one in a uint16.

MapPersistentState and its three subclasses took the id as uint16 while
storing it in a uint32 member, so AddPersistentState handed
WorldPersistentState the deck's 1181646 and it became 1998. That is not a
cosmetic wrap. GetMapEntry() then looks 1998 up in Map.dbc, finds nothing
because the minted row was registered under 1181646, and returns NULL --
which SaveCreatureRespawnTime dereferences unguarded on the very next line.
The first creature to die on a hull takes the world server with it. The same
wrap also splits the manager's own bookkeeping: it files the state under the
real id and asks UnloadIfEmpty to remove the wrapped one, so the state never
comes off.

The deref stays unguarded. A null check there would turn the next width bug
into quiet wrong behaviour instead of a stack trace, and this one was already
quiet enough.

DungeonResetEvent::mapid was a uint16 member fed from a uint32 parameter.

CreatureData::mapid was uint16 while the loader reads that column with
GetUInt32 and GameObjectData::mapid has always been uint32. This is why the
crash has not been seen yet: the row wrapped, the map lookup failed, and the
spawn was rejected as "spawned at nonexistent map (Id: 1998)" -- seven of them
in one session's log, the Azuremyst ferry's crew. The mobs never existed to
die. With this and the preceding commit a hull's crew both load and are keyed
correctly.

Not compiled.

+++ src/game/Object/Creature.h
@@ -272,7 +272,11 @@ struct EquipmentInfoRaw
 struct CreatureData
 {
     uint32 id;                                              // entry in creature_template
-    uint16 mapid;
+    /// uint32, as GameObjectData's has always been. The loader reads this column with
+    /// GetUInt32 and a vessel's deck is a map minted above a million, so as a uint16 it
+    /// wrapped and the row was then rejected as "spawned at nonexistent map" -- which is
+    /// why a hull's crew never appeared.
+    uint32 mapid;
     uint32 modelid_override;                                // overrides any model defined in creature_template
     int32 equipmentId;
     float posX;
